+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/enieber/projects/live-react-elliot/src/index.js":"1","/home/enieber/projects/live-react-elliot/src/App.js":"2","/home/enieber/projects/live-react-elliot/src/serviceWorker.js":"3","/home/enieber/projects/live-react-elliot/src/ClientList.js":"4"},{"size":452,"mtime":1609160164662,"results":"5","hashOfConfig":"6"},{"size":256,"mtime":1609160639696,"results":"7","hashOfConfig":"6"},{"size":4951,"mtime":1609160164662,"results":"8","hashOfConfig":"6"},{"size":4564,"mtime":1609160728867,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"qnp5b6",{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/enieber/projects/live-react-elliot/src/index.js",[],["19","20"],"/home/enieber/projects/live-react-elliot/src/App.js",[],"/home/enieber/projects/live-react-elliot/src/serviceWorker.js",[],"/home/enieber/projects/live-react-elliot/src/ClientList.js",[],{"ruleId":"21","replacedBy":"22"},{"ruleId":"23","replacedBy":"24"},"no-native-reassign",["25"],"no-negated-in-lhs",["26"],"no-global-assign","no-unsafe-negation"]
+[{"/home/enieber/projects/live-react-elliot/src/index.js":"1","/home/enieber/projects/live-react-elliot/src/App.js":"2","/home/enieber/projects/live-react-elliot/src/serviceWorker.js":"3","/home/enieber/projects/live-react-elliot/src/ClientList.js":"4","/home/enieber/projects/live-react-elliot/src/ClientItem.js":"5"},{"size":452,"mtime":1609160164662,"results":"6","hashOfConfig":"7"},{"size":256,"mtime":1609160639696,"results":"8","hashOfConfig":"7"},{"size":4951,"mtime":1609160164662,"results":"9","hashOfConfig":"7"},{"size":983,"mtime":1609161696263,"results":"10","hashOfConfig":"7"},{"size":564,"mtime":1609161528601,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"qnp5b6",{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/enieber/projects/live-react-elliot/src/index.js",[],["23","24"],"/home/enieber/projects/live-react-elliot/src/App.js",[],"/home/enieber/projects/live-react-elliot/src/serviceWorker.js",[],"/home/enieber/projects/live-react-elliot/src/ClientList.js",[],"/home/enieber/projects/live-react-elliot/src/ClientItem.js",[],{"ruleId":"25","replacedBy":"26"},{"ruleId":"27","replacedBy":"28"},"no-native-reassign",["29"],"no-negated-in-lhs",["30"],"no-global-assign","no-unsafe-negation"]
